+++ example/prismapydantic/common/utils.py
@@ -1,0 +1,5 @@
+import os
+
+
+def get_env() -> str:
+    return os.environ.get('ENV', 'easysamdev')